Refactor taxonomy post type badges into hook
