+++ packages/js/src/settings/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useFormikError } from "./use-formik-error";
 export { default as useRouterScrollRestore } from "./use-router-scroll-restore";
+export { default as useTaxonomyPostTypeBadges } from "./use-taxonomy-post-type-badges";
